+++ src/main/java/com/spring/admin/space/controller/SpaceController.java
@@ -1,4 +1,4 @@
-package com.spring.client.space.controller;
+package com.spring.admin.space.controller;
 
 import java.util.List;
 
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.spring.admin.space.domain.Space;
@@ -23,7 +24,6 @@ public class SpaceController {
 	@Setter(onMethod_ = @Autowired)
 	private SpaceService spaceService;
 
-	// 공간 리스트 조회
 	@GetMapping("/spaceList")
 	public String spaceList(Space space, Model model) {
 		List<Space> spaceList= spaceService.spaceList(space);
@@ -32,7 +32,17 @@ public class SpaceController {
 		return "client/space/spaceList";
 	}
 	
-	// 공간 상세 조회
+	@GetMapping("/insertForm")
+	public String insertForm(Space space) {
+		return "client/space/insertForm";
+	}
+	
+	@PostMapping("/spaceInsert")
+	public String spaceInsert(Space space) {
+		spaceService.spaceInsert(space);
+		return "redirect:/space/spaceList";
+	}
+	
 	@GetMapping("/{spNo}")
 	public String spaceDetail(@PathVariable Long spNo, Space space, Model model) {
 		space.setSpNo(spNo);
@@ -44,4 +54,23 @@ public class SpaceController {
 		
 		return "client/space/spaceDetail";
 	}
+
+	@GetMapping("/updateForm")
+	public String updateForm(Space space, Model model) {
+		Space updateData = spaceService.getSpace(space.getSpNo());
+		model.addAttribute("updateData", updateData);
+		return "client/space/updateForm";
+	}
+	
+	@PostMapping("/spaceUpdate")
+	public String spaceUpdate(Space space) {
+		spaceService.spaceUpdate(space);
+		return "redirect:/space/" + space.getSpNo();
+	}
+	
+	@PostMapping("/spaceDelete")
+	public String spaceDelete(Space space) {
+		spaceService.spaceDelete(space);
+		return "redirect:/space/spaceList";
+	}	
 }
