+++ src/main/java/com/spring/admin/space/repository/SpaceDetailRepository.java
@@ -1,0 +1,5 @@
+package com.spring.admin.space.repository;
+
+public class SpaceDetailRepository {
+
+}
