--- src/main/resources/static/css/admin/adminNav.css
@@ -1,37 +1,0 @@
-/* adminNav.css */
-.nav-container {
-    width: 250px;
-    background-color: #343a40;
-    color: #ffffff;
-    height: 100vh;
-}
-
-.sidebar {
-    display: flex;
-    flex-direction: column;
-    padding: 20px;
-}
-
-.sidebar h2 {
-    font-size: 1.5em;
-    margin-bottom: 20px;
-}
-
-.sidebar ul {
-    list-style: none;
-    padding: 0;
-}
-
-.sidebar ul li {
-    margin-bottom: 10px;
-}
-
-.sidebar ul li a {
-    color: #ffffff;
-    text-decoration: none;
-    font-size: 1.2em;
-}
-
-.sidebar ul li a:hover {
-    text-decoration: underline;
-}
