--- src/main/resources/static/css/admin/space/adminSpaceList.css
@@ -1,1 +1,0 @@
-@charset "UTF-8";
