--- src/main/resources/static/css/client/space/spaceDetail.css
@@ -1,17 +1,0 @@
-body {
-    background-color: #f8f9fa;
-}
-
-.card {
-    border-radius: 0.5rem;
-}
-
-.card-img-top {
-    border-top-left-radius: 0.5rem;
-    border-top-right-radius: 0.5rem;
-}
-
-.btn-primary {
-    background-color: #007bff;
-    border-color: #007bff;
-}
