--- src/test/java/com/spring/admin/space/domain/SpaceDetailRepository.java
@@ -1,5 +1,0 @@
-package com.spring.admin.space.domain;
-
-public class SpaceDetailRepository {
-
-}
